Restructure
Rename 'code' folder to 'src', since a folder named code can not be a package (reserved package name).
Move main files in code to folder called projects and refactor filepaths and imports.
