+++ src/scripts/reset_store.sh
